// front/truespotFront/src/model/climbingRoad.ts
/**
 * Swagger Maven Plugin Sample
 * This is a sample.
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {Area} from './area';


export interface ClimbingRoad {
    id?: number;
    area?: Area;
    name?: string;
    description?: string;
    type?: string;
    number?: number;
    letter?: string;
    symbol?: string;
}